refactor(ui): rehome history utility
