+++ npm/ui/core/src/families/interaction/history/history-stack.ts
